// types.ts
export enum TaskDifficulty {
  EASY = 'Easy',
  MEDIUM = 'Medium',
  HARD = 'Hard',
  SAVAGE = 'Savage',
}
// ...
export interface SideQuest {
  id: string;
  description: string;
  difficulty: TaskDifficulty;
  dailyGoal: number; // How many times it can be done per day. 0 for infinite.
  completions: { [date: string]: number }; // Key is YYYY-MM-DD, value is count for that day
}